shared_mem: build DAG — generated histogram kernel passes C-sim (milestone)
The GENERATED kernel (gen/hist.cpp from HistAccel) compiles with the hand-written
datapath hooks and passes Vitis C-sim against the HistogramAccel golden — the
milestone the increment toy was built to de-risk, now on the real example.

- shared_mem_build.py: generate_vitis_sources (include headers via
  HistTest.gen_vitis_code + MemMgrStep, then gen/hist.{cpp,hpp}) and HistCase
  (per-case inputs + golden status/counts). run_gen.tcl compiles the generated
  kernel + hooks; hist_csim_tb.cpp is a robust csim driver (clamped allocs,
  counts read back only on NO_ERROR) so it can drive validation-failure cases as
  well as the happy path.
- The generated header is now a TB drop-in: alongside the kernel's buffer-bound
  constants it emits the total m_axi region (max_mem_words), the interface widths
  (stream_dwidth/mem_dwidth/mem_awidth), and the axis_word_t/mem_word_t typedefs
  the testbench needs (guarded by m_axi presence — poly/regmap unaffected). Added
  a mem_awidth HwParam to HistAccel for the address width.
- The hand-written validate/respond hooks convert between the scoped HistError
  enum (HistResp.status) and the ap_uint<8> the codegen carries the status as.

tests/examples/test_hist_csim.py (vitis-gated, no soft-skip) runs C-sim across
nbins==1 (the unconditional zero-count edges read), nbins>1 with 6 and 12 bins,
and a validation-failure case, asserting status + counts match the golden.
Empirically green on Vitis HLS 2025.1: all four cases pass.

The hand-written hist.cpp/hist_tb.cpp stay the diff targets (retired in Phase 6).

+++ examples/shared_mem/hist_validate_impl.cpp
@@ -15,16 +15,18 @@ ap_uint<8> validate(HistCmd cmd) {
     const int ndata = (int)cmd.ndata;
     const int nbins = (int)cmd.nbins;
 
+    // HistError is a scoped enum; the codegen carries the status as ap_uint<8>
+    // (matching the generated hook signature), so convert via the integer value.
     if (ndata <= 0 || ndata > max_ndata)
-        return (ap_uint<8>)HistError::INVALID_NDATA;
+        return (ap_uint<8>)static_cast<unsigned int>(HistError::INVALID_NDATA);
     if (nbins <= 0 || nbins > max_nbins)
-        return (ap_uint<8>)HistError::INVALID_NBINS;
+        return (ap_uint<8>)static_cast<unsigned int>(HistError::INVALID_NBINS);
     if (!memmgr::is_word_aligned<32>(cmd.data_addr) ||
         !memmgr::is_word_aligned<32>(cmd.bin_edges_addr) ||
         !memmgr::is_word_aligned<32>(cmd.cnt_addr))
-        return (ap_uint<8>)HistError::ADDRESS_ERROR;
+        return (ap_uint<8>)static_cast<unsigned int>(HistError::ADDRESS_ERROR);
 
-    return (ap_uint<8>)HistError::NO_ERROR;
+    return (ap_uint<8>)static_cast<unsigned int>(HistError::NO_ERROR);
 }
 
 }  // namespace hist_impl
